links to external files in regular field text
- at the moment the prefix is ____ (4*underscore)and the
page separator is ___ (3*underscore).
- E.g. you have a file "France.pdf" in a relative folder,
when you have ____france___3 anywhere in your note
this is transformed into a hyperlink in the reviewer
and you can doubleclick it in the editor to open
the external file. In the reviewer ____France.pdf___3
is also transformed into a clickable hyperlink.
- I chose ____ and ___ because when you double-click
them the whole string is selected (whereas this
doesn't work for characters like "-", "#", etc.)
- double-clicking stops at a ".". So you can
never select a file and the extension by double
clicking it. That's why I added a function to guess
the extension so that you can use
 ____france___3.
- the files that start with linked__ are copied over
from my add-on "link Cards Notes and Preview them
in extra window" and slightly adjusted.